Reduce submit transformer complexity
