+++ src/applications/disability-benefits/all-claims/utils/submit.js
@@ -2,10 +2,13 @@ import _ from 'platform/utilities/data';
 import {
   PTSD_INCIDENT_ITERATION,
   PTSD_CHANGE_LABELS,
+  ATTACHMENT_KEYS,
+  causeTypes,
   disabilityActionTypes,
 } from '../constants';
 
 import {
+  isDisabilityPtsd,
   disabilityIsSelected,
   hasGuardOrReservePeriod,
   sippableId,
@@ -297,4 +300,127 @@ export const cleanUpMailingAddress = formData => {
     {},
   );
   return { ...formData, mailingAddress };
+};
+
+// Add 'cause' of 'NEW' to new ptsd disabilities since form does not ask
+export const addPTSDCause = formData =>
+  formData.newDisabilities
+    ? _.set(
+        'newDisabilities',
+        formData.newDisabilities.map(
+          disability =>
+            isDisabilityPtsd(disability.condition)
+              ? _.set('cause', causeTypes.NEW, disability)
+              : disability,
+        ),
+        formData,
+      )
+    : formData;
+
+export const addForm4142 = formData => {
+  if (!formData.providerFacility) {
+    return formData;
+  }
+  const clonedData = _.cloneDeep(formData);
+  clonedData.form4142 = {
+    ...(clonedData.limitedConsent && {
+      limitedConsent: clonedData.limitedConsent,
+    }),
+    ...(clonedData.providerFacility && {
+      providerFacility: transformProviderFacilities(
+        clonedData.providerFacility,
+      ),
+    }),
+  };
+  delete clonedData.limitedConsent;
+  delete clonedData.providerFacility;
+  return clonedData;
+};
+
+export const addForm0781 = formData => {
+  const clonedData = _.cloneDeep(formData);
+  const incidentKeys = getFlatIncidentKeys();
+  const incidents = incidentKeys
+    .filter(incidentKey => clonedData[incidentKey])
+    .map(incidentKey => ({
+      ...cleanUpPersonsInvolved(clonedData[incidentKey]),
+      personalAssault: incidentKey.includes('secondary'),
+    }));
+  incidentKeys.forEach(incidentKey => {
+    delete clonedData[incidentKey];
+  });
+  if (incidents.length > 0) {
+    clonedData.form0781 = {
+      incidents,
+      remarks: clonedData.additionalRemarks781,
+      additionalIncidentText: clonedData.additionalIncidentText,
+      additionalSecondaryIncidentText:
+        clonedData.additionalSecondaryIncidentText,
+      otherInformation: [
+        ...getPtsdChangeText(clonedData.physicalChanges),
+        _.get('physicalChanges.otherExplanation', clonedData, ''),
+        ...getPtsdChangeText(clonedData.socialBehaviorChanges),
+        _.get('socialBehaviorChanges.otherExplanation', clonedData, ''),
+        ...getPtsdChangeText(clonedData.mentalChanges),
+        _.get('mentalChanges.otherExplanation', clonedData, ''),
+        ...getPtsdChangeText(clonedData.workBehaviorChanges),
+        _.get('workBehaviorChanges.otherExplanation', clonedData, ''),
+        _.get('additionalChanges', clonedData, ''),
+      ].filter(info => info.length > 0),
+    };
+    delete clonedData.physicalChanges;
+    delete clonedData.socialBehaviorChanges;
+    delete clonedData.mentalChanges;
+    delete clonedData.workBehaviorChanges;
+    delete clonedData.additionalChanges;
+    delete clonedData.additionalRemarks781;
+    delete clonedData.additionalIncidentText;
+    delete clonedData.additionalSecondaryIncidentText;
+  }
+  return clonedData;
+};
+
+export const addForm8940 = formData => {
+  const clonedData = _.cloneDeep(formData);
+  const { unemployability } = clonedData;
+
+  if (unemployability) {
+    const disabilities = getDisabilities(formData);
+
+    clonedData.form8940 = {
+      unemployability: {
+        ...unemployability,
+        disabilityPreventingEmployment: disabilities
+          .filter(disability => disability.unemployabilityDisability)
+          .map(disability => getDisabilityName(disability))
+          .join(),
+        underDoctorHopitalCarePast12M:
+          unemployability.underDoctorsCare || unemployability.hospitalized,
+        mostEarningsInAYear: (
+          unemployability.mostEarningsInAYear || ''
+        ).toString(),
+        disabilityPreventMilitaryDuties:
+          unemployability.disabilityPreventMilitaryDuties === 'yes',
+      },
+    };
+
+    delete clonedData.form8940.unemployability.underDoctorsCare;
+    delete clonedData.form8940.unemployability.hospitalized;
+    delete clonedData.unemployability;
+  }
+
+  return clonedData;
+};
+
+// Flatten all attachment pages into attachments ARRAY
+export const addFileAttachments = formData => {
+  const clonedData = _.cloneDeep(formData);
+  let attachments = [];
+
+  ATTACHMENT_KEYS.forEach(key => {
+    const documentArr = _.get(key, clonedData, []);
+    attachments = [...attachments, ...documentArr];
+    delete clonedData[key];
+  });
+  return { ...clonedData, ...(attachments.length && { attachments }) };
 };
